Revert 58f2b32..598ff6c
This rolls back to commit 58f2b3217eecb7d9b3b94896686683105a91b488.
