--- ProjectSettings/haxe-luxe-settings.sublime-settings
@@ -1,16 +1,0 @@
-{
-   "name": "luxe",
-   "extension" : "luxe.xml",
-   "run":["haxelib","run","lime"],
-   "build_targets": {
-        "HTML5 - test"                    : ["test", "html5", "-v"],
-        "HTML5 - debug"                   : ["test", "html5", "-debug", "-v"],
-        "C++ - test"                      : ["test", "cpp", "-v"],
-        "C++ - debug"                     : ["test", "cpp", "-debug", "-v"],
-        "Android - test"                  : ["test", "android", "-v"],
-        "Android - debug"                 : ["test", "android","-debug", "-v"],
-        "iOS - test in iPhone simulator"  : ["test", "ios","-simulator", "-v"],
-        "iOS - test in iPad simulator"    : ["test", "ios", "-simulator","-ipad", "-v"],
-        "iOS - update XCode project"      : ["update", "ios", "-v"],
-    }   
-}
