--- ProjectSettings/haxe-openfl-settings.sublime-settings
@@ -1,32 +1,0 @@
-{
-   "name": "openfl",
-   "extension" : "xml",
-   "run":["haxelib","run","openfl"],
-   "build_targets": {
-        "Flash - test"                      : ["test",  "flash","-debug"],
-        "Flash - build only"                : ["build", "flash","-debug"],
-        "HTML5 - test"                      : ["test",  "html5","-debug"],
-        "HTML5 - build only"                : ["build", "html5","-debug"],
-        "C++ - test"                        : ["test",  "cpp","-debug"],
-        "C++ - build only"                  : ["build", "cpp","-debug"],
-        "Linux - test"                      : ["test",  "linux","-debug"],
-        "Linux - build only"                : ["build", "linux","-debug"], 
-        "Linux 64 - test"                   : ["test",  "linux","-64","-debug"],
-        "Linux 64 - build only"             : ["build", "linux","-64","-debug"],
-        "iOS - test in iPhone simulator"    : ["test",  "ios","-simulator","-debug"],
-        "iOS - test in iPad simulator"      : ["test",  "ios","-simulator","-ipad","-debug"],
-        "iOS - update XCode project"        : ["update","ios","-debug"],
-        "Android - test"                    : ["test",  "android","-debug"],
-        "Android - build only"              : ["build", "android","-debug"],
-        "WebOS - test"                      : ["test",  "webos","-debug"],
-        "WebOS - build only"                : ["build", "webos","-debug"],
-        "Neko - test"                       : ["test",  "neko","-debug"],
-        "Neko - build only"                 : ["build", "neko","-debug"],
-        "Neko 64 - test"                    : ["test",  "neko","-64","-debug"],
-        "Neko 64 - build only"              : ["build", "neko","-64","-debug"],
-        "BlackBerry - test"                 : ["test",  "blackberry","-debug"],
-        "BlackBerry - build only"           : ["build", "blackberry","-debug"],
-        "Emscripten - test"                 : ["test",  "emscripten","-debug"],
-        "Emscripten - build only"           : ["build", "emscripten","-debug"]
-    } 
-}
