+++ Support/Side Bar.sublime-menu
@@ -5,10 +5,10 @@
     "children":  
     [  
         { 
-            "id" : "haxe-create-class", 
-            "caption" : "Class",
-            "command" : "haxe_create_type",
-            "args" : { "paths": [] , "t" : "class" }
+        	"id" : "haxe-create-class", 
+        	"caption" : "Class",
+        	"command" : "haxe_create_type",
+        	"args" : { "paths": [] , "t" : "class" }
         },
         { 
             "id" : "haxe-create-interface", 
@@ -28,38 +28,6 @@
             "command" : "haxe_create_type",
             "args" : { "paths": [] , "t" : "typedef" }
         }
-    ]  
-}, 
-{  
-    "id": "haxe-select-project-file",  
-    "caption" : "Use Project as",
-    "children":  
-    [  
-        { 
-        	"id" : "haxe-select-openfl", 
-        	"caption" : "OpenFL (*.xml)",
-        	"command" : "haxe_select_project",
-        	"args" : { "paths": [] , "t" : "class" }
-        },
-        { 
-            "id" : "haxe-select-lime", 
-            "caption" : "Lime (*.lime.xml)",
-            "command" : "haxe_select_project",
-            "args" : { "paths": [] , "t" : "interface" }
-        },
-        { 
-            "id" : "haxe-select-luxe", 
-            "caption" : "Luxe (*.luxe.xml)",
-            "command" : "haxe_select_project",
-            "args" : { "paths": [] , "t" : "interface" }
-        },
-        { 
-            "id" : "haxe-select-haxe", 
-            "caption" : "Haxe (*.hxml)",
-            "command" : "haxe_select_project",
-            "args" : { "paths": [] , "t" : "enum" }
-        }
-        
     ]  
 }  
 ]
