--- features/__init__.py
@@ -1,4 +1,0 @@
-__version__ = "1.0.0"
-__version_info__ = (1, 0, 0)
-
-from .haxelib import *
