--- features/haxelib/__init__.py
@@ -1,20 +1,0 @@
-
-import sublime
-
-__version__ = "1.0.0"
-__version_info__ = (1, 0, 0)
-
-print("Haxe : Reloading haxelib module")
-
-from .haxelib_list_installed import HaxelibListInstalled
-from .haxelib_install_lib import HaxelibInstallLib
-from .haxelib_list_libs import HaxelibListLibs
-from .haxelib_upgrade_libs import HaxelibUpgradeLibs
-
-__all__ = [
-    'HaxelibListInstalled',
-    'HaxelibInstallLib',
-    'HaxelibListLibs',
-    'HaxelibUpgradeLibs'
-]
-
